Island: Move jwt_required to monkey_island.cc.flask_utils
