+++ monkey/monkey_island/cc/flask_utils/__init__.py
@@ -1,2 +1,3 @@
 from .abstract_resource import AbstractResource
 from .flask_di_wrapper import FlaskDIWrapper
+from .jwt_required import jwt_required
