+++ monkey/monkey_island/cc/resources/agent_signals/terminate_all_agents.py
@@ -5,9 +5,8 @@ from json import JSONDecodeError
 from flask import request
 
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
-from monkey_island.cc.flask_utils import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 from monkey_island.cc.models import TerminateAllAgents as TerminateAllAgentsObject
-from monkey_island.cc.resources.request_authentication import jwt_required
 
 logger = logging.getLogger(__name__)
