+++ monkey/monkey_island/cc/resources/island_mode.py
@@ -5,10 +5,9 @@ from http import HTTPStatus
 from flask import request
 
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
-from monkey_island.cc.flask_utils import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 from monkey_island.cc.models import IslandMode as IslandModeEnum
 from monkey_island.cc.repositories import ISimulationRepository
-from monkey_island.cc.resources.request_authentication import jwt_required
 
 logger = logging.getLogger(__name__)
