+++ monkey/monkey_island/cc/resources/remote_run.py
@@ -4,8 +4,7 @@ from typing import Sequence
 from botocore.exceptions import ClientError, NoCredentialsError
 from flask import jsonify, make_response, request
 
-from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.resources.request_authentication import jwt_required
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 from monkey_island.cc.services import AWSService
 from monkey_island.cc.services.aws import AWSCommandResults
