+++ monkey/monkey_island/cc/resources/report_generation_status.py
@@ -1,8 +1,7 @@
 from flask import jsonify
 
-from monkey_island.cc.flask_utils import AbstractResource
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 from monkey_island.cc.repositories import IAgentRepository
-from monkey_island.cc.resources.request_authentication import jwt_required
 from monkey_island.cc.services.infection_lifecycle import is_report_done
