+++ monkey/monkey_island/cc/resources/reset_agent_configuration.py
@@ -3,8 +3,7 @@ from http import HTTPStatus
 from flask import make_response
 
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
-from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.resources.request_authentication import jwt_required
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 
 
 class ResetAgentConfiguration(AbstractResource):
