+++ monkey/monkey_island/cc/resources/security_report.py
@@ -1,5 +1,4 @@
-from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.resources.request_authentication import jwt_required
+from monkey_island.cc.flask_utils import AbstractResource, jwt_required
 from monkey_island.cc.services.reporting.report import ReportService
